Nur noch Long und kein Integer beim Buchen benutzen

+++ aero.minova.core.application.system.service/src/main/java/aero/minova/core/application/system/covid/test/print/domain/MeetingFormInformation.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @EqualsAndHashCode
 @NoArgsConstructor
 public class MeetingFormInformation {
-	private int testStreckKeyLong;
+	private long testStreckKeyLong;
 	private String starttime;
 	private long testPersonKeyLong;
 }
